add new document button to content list
