+++ src/components/content-types/ContentList.css
@@ -6,9 +6,25 @@
 .ContentListItem--insertTop:after {
   content: "";
   position: absolute;
-  top:0;
+  top: 0;
   left: 0;
   right: 0;
   height: 1px;
   background: var(--colorSecondaryGrey);
+}
+
+.ContentList--newItem {
+  margin-top: 16px;
+  padding: 4px 8px;
+  font-size: 1.1em;
+  font-weight: 500;
+  color: var(--colorSecondaryGrey);
+  display: flex;
+  gap: 8px;
+  align-items: center;
+  cursor: pointer;
+}
+
+.ContentList--newItem-icon {
+  padding: 0 6px;
 }
