+++ src/components/content-types/workspace/TitleBar.css
@@ -47,6 +47,7 @@ button[disabled].TitleBar-menuItem .fa {
   border: none;
   outline: none;
   position: relative;
+  cursor: pointer;
 }
 
 .TitleBar-menuItem:active,
